implemented abstraction to planner interface code
