+++ rosplan_planning_system/include/rosplan_planning_system/PlannerInterface/FDPlannerInterface.h
@@ -19,9 +19,6 @@ namespace KCL_rosplan {
 	{
 	private:
 
-		/* runs external commands */
-		std::string runCommand(std::string cmd);
-
 	protected:
 
 		bool runPlanner();
