+++ rosplan_planning_system/include/rosplan_planning_system/PlannerInterface/LPGPlannerInterface.h
@@ -18,9 +18,6 @@ namespace KCL_rosplan {
 	{
 	private:
 
-		/* runs external commands */
-		std::string runCommand(std::string cmd);
-
 	protected:
 
 		bool runPlanner();
