+++ rosplan_planning_system/src/PlannerInterface/FFPlannerInterface.cpp
@@ -6,21 +6,8 @@ namespace KCL_rosplan {
 	/* constructor */
 	/*-------------*/
 
-	FFPlannerInterface::FFPlannerInterface(ros::NodeHandle& nh)
-	{
-		node_handle = &nh;
-
-		plan_server = new actionlib::SimpleActionServer<rosplan_dispatch_msgs::PlanAction>((*node_handle), "start_planning", boost::bind(&PlannerInterface::runPlanningServerAction, this, _1), false);
-
-		// publishing raw planner output
-		std::string plannerTopic = "planner_output";
-		node_handle->getParam("planner_topic", plannerTopic);
-		plan_publisher = node_handle->advertise<std_msgs::String>(plannerTopic, 1, true);
-
+	FFPlannerInterface::FFPlannerInterface(ros::NodeHandle& nh): PlannerInterface(nh) {
 		node_handle->param<bool>("use_ffha", this->use_ffha, false);
-
-		// start planning action server
-		plan_server->start();
 	}
 
 	FFPlannerInterface::~FFPlannerInterface()
@@ -28,19 +15,6 @@ namespace KCL_rosplan {
 		delete plan_server;
 	}
 
-	/**
-	 * Runs external commands
-	 */
-	std::string FFPlannerInterface::runCommand(std::string cmd) {
-		std::string data;
-		FILE *stream;
-		char buffer[1000];
-		stream = popen(cmd.c_str(), "r");
-		while ( fgets(buffer, 1000, stream) != NULL )
-			data.append(buffer);
-		pclose(stream);
-		return data;
-	}
 
 	/*------------------*/
 	/* Plan and process */
@@ -52,26 +26,13 @@ namespace KCL_rosplan {
 	bool FFPlannerInterface::runPlanner() {
 		planner_output.clear(); // Clear previous plan
 		// save problem to file for planner
-		if(use_problem_topic && problem_instance_received) {
-			ROS_INFO("KCL: (%s) (%s) Writing problem to file.", ros::this_node::getName().c_str(), problem_name.c_str());
-			std::ofstream dest;
-			dest.open((problem_path).c_str());
-			dest << problem_instance;
-			dest.close();
-		}
+		saveProblem();
 
 		// prepare the planner command line
-		std::string str = planner_command;
-		std::size_t dit = str.find("DOMAIN");
-		if(dit!=std::string::npos) str.replace(dit,6,domain_path);
-		std::size_t pit = str.find("PROBLEM");
-		if(pit!=std::string::npos) str.replace(pit,7,problem_path);
-		std::string commandString = str + " > " + data_path + "plan.pddl";
+		std::string commandString = prepareCommand() + " > " + data_path + "plan.pddl";
 
 		// call the planer
-		ROS_INFO("KCL: (%s) (%s) Running: %s", ros::this_node::getName().c_str(), problem_name.c_str(),  commandString.c_str());
-		std::string plan = runCommand(commandString.c_str());
-		ROS_INFO("KCL: (%s) (%s) Planning complete", ros::this_node::getName().c_str(), problem_name.c_str());
+		callPlanner(commandString);
 
 		// check the planner solved the problem
 		std::ifstream planfile;
@@ -81,6 +42,7 @@ namespace KCL_rosplan {
 		bool solved = false;
 		while (not solved and std::getline(planfile, line)) {
 			// skip lines until there is a plan
+			std::cout << line << std::endl;
 			if (line.find("ff: found legal plan") != line.npos) {
 				solved = true;
 			}
@@ -125,38 +87,9 @@ namespace KCL_rosplan {
 		}
 		planfile.close();
 
-		if(!solved) ROS_INFO("KCL: (%s) (%s) Plan was unsolvable.", ros::this_node::getName().c_str(), problem_name.c_str());
-		else ROS_INFO("KCL: (%s) (%s) Plan was solved.", ros::this_node::getName().c_str(), problem_name.c_str());
+		solvedMessages(solved);
 
 		return solved;
 	}
 
 } // close namespace
-
-	/*-------------*/
-	/* Main method */
-	/*-------------*/
-
-	int main(int argc, char **argv) {
-
-		srand (static_cast <unsigned> (time(0)));
-
-		ros::init(argc,argv,"rosplan_planner_interface");
-		ros::NodeHandle nh("~");
-
-		KCL_rosplan::FFPlannerInterface pi(nh);
-
-		// subscribe to problem instance
-		std::string problemTopic = "problem_instance";
-		nh.getParam("problem_topic", problemTopic);
-		ros::Subscriber problem_sub = nh.subscribe(problemTopic, 1, &KCL_rosplan::PlannerInterface::problemCallback, dynamic_cast<KCL_rosplan::PlannerInterface*>(&pi));
-
-		// start the planning services
-		ros::ServiceServer service1 = nh.advertiseService("planning_server", &KCL_rosplan::PlannerInterface::runPlanningServerDefault, dynamic_cast<KCL_rosplan::PlannerInterface*>(&pi));
-		ros::ServiceServer service2 = nh.advertiseService("planning_server_params", &KCL_rosplan::PlannerInterface::runPlanningServerParams, dynamic_cast<KCL_rosplan::PlannerInterface*>(&pi));
-
-		ROS_INFO("KCL: (%s) Ready to receive", ros::this_node::getName().c_str());
-		ros::spin();
-
-		return 0;
-	}
